pyproject.toml
[project]
name = "delivery-2-0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-wtf>=1.2.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
